refactor posts list to use view model
