+++ apps/frontend/src/features/posts/components/PostList.vue
@@ -1,18 +1,10 @@
 <script setup lang="ts">
-import { ref, computed, onMounted, watch } from 'vue'
-import { usePostStore } from '../stores/postStore'
-import { useI18n } from 'vue-i18n'
 import PostCard from './PostCardPostIt.vue'
 import PostEdit from './PostEdit.vue'
 import PostFullView from './PostFullView.vue'
-import type {
-  PublicPostWithProfile,
-  OwnerPost,
-  PostQueryInput,
-  NearbyPostQueryInput,
-} from '@zod/post/post.dto'
 import { type PostTypeType } from '@zod/generated'
-import ViewModeToggler from '@/features/shared/ui/ViewModeToggler.vue'
+
+import { usePostListViewModel } from '../composables/usePostListViewModel'
 
 interface Props {
   title?: string
@@ -30,136 +22,25 @@ const props = withDefaults(defineProps<Props>(), {
   emptyMessage: 'No posts found',
 })
 
-const { t } = useI18n()
-const postStore = usePostStore()
-
-const selectedType = ref<string>(props.type || '')
-const currentPage = ref(0)
-const pageSize = 20
-const showFullView = ref(false)
-const showEditModal = ref(false)
-const selectedPost = ref<PublicPostWithProfile | OwnerPost | null>(null)
-const editingPost = ref<OwnerPost | null>(null)
-
-const posts = computed(() => {
-  if (props.scope === 'my') {
-    return postStore.myPosts
-  }
-  return postStore.posts
-})
-
-const canLoadMore = computed(() => {
-  return posts.value.length >= (currentPage.value + 1) * pageSize
-})
-
-const buildQuery = (): PostQueryInput => ({
-  type: (selectedType.value as PostTypeType) || undefined,
-  limit: pageSize,
-  offset: currentPage.value * pageSize,
-})
-
-const buildNearbyQuery = (): NearbyPostQueryInput => ({
-  ...buildQuery(),
-  lat: props.nearbyParams!.lat,
-  lon: props.nearbyParams!.lon,
-  radius: props.nearbyParams!.radius || 50,
-})
-
-const loadPosts = async (append = false) => {
-  if (!append) {
-    currentPage.value = 0
-  }
-
-  const query = buildQuery()
-
-  switch (props.scope) {
-    case 'nearby':
-      if (props.nearbyParams) {
-        await postStore.fetchNearbyPosts(buildNearbyQuery())
-      }
-      break
-    case 'recent':
-      await postStore.fetchRecentPosts(query)
-      break
-    case 'my':
-      await postStore.fetchMyPosts(query)
-      break
-    default:
-      await postStore.fetchPosts(query)
-  }
-}
-
-const handleTypeFilter = () => {
-  loadPosts()
-}
-
-const handleLoadMore = () => {
-  currentPage.value++
-  loadPosts(true)
-}
-
-const handleRetry = () => {
-  postStore.clearError()
-  loadPosts()
-}
-
-const handlePostClick = (post: PublicPostWithProfile | OwnerPost) => {
-  selectedPost.value = post
-  showFullView.value = true
-}
-
-const handlePostEdit = (post: PublicPostWithProfile | OwnerPost) => {
-  editingPost.value = post as OwnerPost
-  showEditModal.value = true
-}
-
-const handlePostDelete = async (post: PublicPostWithProfile | OwnerPost) => {
-  if (confirm(t('posts.confirm_delete'))) {
-    const success = await postStore.deletePost(post.id)
-    if (success) {
-      closeFullView()
-    }
-  }
-}
-
-const handlePostSaved = (post: OwnerPost) => {
-  closeEditModal()
-  // Refresh the list to show updated post
-  loadPosts()
-}
-
-const closeFullView = () => {
-  showFullView.value = false
-  selectedPost.value = null
-}
-
-const closeEditModal = () => {
-  showEditModal.value = false
-  editingPost.value = null
-}
-
-// Watch for prop changes
-watch(
-  () => props.type,
-  newType => {
-    selectedType.value = newType || ''
-    loadPosts()
-  }
-)
-
-watch(
-  () => props.nearbyParams,
-  () => {
-    if (props.scope === 'nearby') {
-      loadPosts()
-    }
-  },
-  { deep: true }
-)
-
-onMounted(() => {
-  loadPosts()
-})
+const {
+  postStore,
+  posts,
+  selectedType,
+  canLoadMore,
+  showFullView,
+  showEditModal,
+  selectedPost,
+  editingPost,
+  handleTypeFilter,
+  handleLoadMore,
+  handleRetry,
+  handlePostClick,
+  handlePostEdit,
+  handlePostDelete,
+  handlePostSaved,
+  closeFullView,
+  closeEditModal,
+} = usePostListViewModel(props)
 </script>
 
 <template>
